ZynAddSubFX/FLTK: updated to SVN revision 8805
Updated FLTK to SVN revision 8805 of branch-1.3.

+++ plugins/zynaddsubfx/fltk/src/fl_boxtype.cxx
@@ -1,9 +1,9 @@
 //
-// "$Id: fl_boxtype.cxx 7469 2010-04-07 23:17:33Z matt $"
+// "$Id: fl_boxtype.cxx 7903 2010-11-28 21:06:39Z matt $"
 //
 // Box drawing code for the Fast Light Tool Kit (FLTK).
 //
-// Copyright 1998-2009 by Bill Spitzak and others.
+// Copyright 1998-2010 by Bill Spitzak and others.
 //
 // This library is free software; you can redistribute it and/or
 // modify it under the terms of the GNU Library General Public
@@ -81,19 +81,19 @@ void fl_frame(const char* s, int x, int y, int w, int h) {
   uchar *g = fl_gray_ramp();
   if (h > 0 && w > 0) for (;*s;) {
     // draw top line:
-    fl_color(g[*s++]);
+    fl_color(g[(int)*s++]);
     fl_xyline(x, y, x+w-1);
     y++; if (--h <= 0) break;
     // draw left line:
-    fl_color(g[*s++]);
+    fl_color(g[(int)*s++]);
     fl_yxline(x, y+h-1, y);
     x++; if (--w <= 0) break;
     // draw bottom line:
-    fl_color(g[*s++]);
+    fl_color(g[(int)*s++]);
     fl_xyline(x, y+h-1, x+w-1);
     if (--h <= 0) break;
     // draw right line:
-    fl_color(g[*s++]);
+    fl_color(g[(int)*s++]);
     fl_yxline(x+w-1, y+h-1, y);
     if (--w <= 0) break;
   }
@@ -115,19 +115,19 @@ void fl_frame2(const char* s, int x, int y, int w, int h) {
   uchar *g = fl_gray_ramp();
   if (h > 0 && w > 0) for (;*s;) {
     // draw bottom line:
-    fl_color(g[*s++]);
+    fl_color(g[(int)*s++]);
     fl_xyline(x, y+h-1, x+w-1);
     if (--h <= 0) break;
     // draw right line:
-    fl_color(g[*s++]);
+    fl_color(g[(int)*s++]);
     fl_yxline(x+w-1, y+h-1, y);
     if (--w <= 0) break;
     // draw top line:
-    fl_color(g[*s++]);
+    fl_color(g[(int)*s++]);
     fl_xyline(x, y, x+w-1);
     y++; if (--h <= 0) break;
     // draw left line:
-    fl_color(g[*s++]);
+    fl_color(g[(int)*s++]);
     fl_yxline(x, y+h-1, y);
     x++; if (--w <= 0) break;
   }
@@ -425,5 +425,5 @@ void Fl_Widget::draw_box(Fl_Boxtype t, int X, int Y, int W, int H, Fl_Color c) c
 }
 
 //
-// End of "$Id: fl_boxtype.cxx 7469 2010-04-07 23:17:33Z matt $".
+// End of "$Id: fl_boxtype.cxx 7903 2010-11-28 21:06:39Z matt $".
 //
